[Fix] Python skew implementation
The Polygon object is allowed to have extra un-indexed vertices (since
it does not necessarily own the vertex array) or internal vertices which
lie on the edge of a hole.
Since we ultimately want to find the convex hull of two skewed polygons,
we need to ensure no extraneous vertices outside of the polygon border
are included when skew-ing. This is easily achieved by using the border
list to index the vertex array.